refactor: TokenParser 클래스 분리
